Move notifySlackUser into slack directory
